Fix Docker image
Although the "executable" script was working fine in normal usage, when
attempting to put this package into a Docker container I was getting
`exec format error`. This especially confused me, because I wasn't
building any binaries. It turns out that the code for the executable
script that I found online put the shebang/interpreter directive on the
second line, and when you `exec` a text file and expect it to be parsed
by the specified interpreter, the shebang must be the first characters
in the file.

This was VERY irritating to track down.